Added userid to users

+++ PMS/model/users.js
@@ -16,16 +16,34 @@ exports.checkLoginDetails = (username, password)=>{
     return check;
 }
 
+function generateRandomId(length) {
+    const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    let randomId = '';
+    for (let i = 0; i < length; i++) {
+        randomId += characters.charAt(Math.floor(Math.random() * characters.length));
+    }
+    return randomId;
+}
+
 exports.signUpUser = (username, password) => {
     try {
+        const randomId = generateRandomId(8);
+        console.log(randomId);
+
         // Check if username already exists
         const userExists = userInfo.some(user => user.username === username);
         if (userExists) {
             throw new Error('Username already exists');
         }
+        const randomIdExists = userInfo.some(randomId => user.randomId === randomId)
+        if(randomIdExists){
+            throw new Error('ID already exists');
+        }
+        
 
         // Add new user to user data
-        userInfo.push({ username, password });
+        userInfo.push({randomId,username, password });
+        console.log("Pushed ", randomId, " ", username, " ", password);
 
         // Write updated user data back to JSON file
         fs.writeFileSync('./userdb.json', JSON.stringify(userInfo));
